+++ CHANGES.md
@@ -1,6 +1,6 @@
 unreleased
 --------------
-- gluten-async: update async and core to v0.17.0
+- gluten-async: Silence async deprecation warnings
   ([#78)(https://github.com/anmonteiro/gluten/pull/78))
 
 0.5.1 2024-06-04